### 2023-10-06 #5928 @editorialbot
[PRE REVIEW]: PyXAB - A Python Library for X-Armed Bandit and Online Blackbox Optimization
**Submitting author:** @WilliamLwj (<a href="http://orcid.org/0000-0003-1872-4595">Wenjie Li</a>)
**Repository:** https://github.com/WilliamLwj/PyXAB
**Branch with paper.md** (empty if default branch): paper
**Version:** v0.3.0
