### 2024-03-19 #5928 @drvinceknight
:wave: @pescap, @Otomisin, @KBodolai  & @holl-, would any of you be willing to review this submission for JOSS? We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: https://joss.readthedocs.io/en/latest/review_criteria.html
