### 2024-05-19 #6507 @Otomisin
@drvinceknight I think I am done with the review,
Regarding, **A statement of need:** Does the paper have a section titled 'Statement of need' that clearly states what problems the software is designed to solve, who the target audience is, and its relation to other work?

_The paper does not have a section explicitly titled "Statement of need," but the content relevant to this purpose is dispersed throughout the introduction and other sections_

**Data sharing: If the paper contains original data, data are accessible to the reviewers. If the paper contains no original data, please check this item.**
_There does not seem to be a need for data sharing, hence data was not provided, there is however a sample data for real life examle [here ](https://pyxab.readthedocs.io/en/latest/getting_started/auto_examples/plot_VHCT_Landmine.html)_

**Functionality: Have the functional claims of the software been confirmed?**
I made a test by recreating with the example (1-D Example) on my personal PC,  from [1-D Example](https://pyxab.readthedocs.io/en/latest/getting_started/auto_examples/plot_HOO_Garland.html )

Please let me know if anything else is needed from my side.
