### 2024-07-08 #6507 @WilliamLwj
> Thanks for your review @Otomisin. @WilliamLwj could you let me know how revisions are going?

Hi @drvinceknight, to be honest, we are a bit unsure what we should do next because we have't heard back for quite long, and this is our first time submitting to JOSS. We thought we should wait for the reviews from @KBodolai to be completed before we start prepare the revisions. Also, could you let us know how to upload the revisions?

In the meantime, we do not see a lot of changes requested by the reviewers. The two (potential) improvements we identify are:

1. Add missing DOIs for the references - We can work on this one if needed.
2. Add a "statement of need" section that clearly states what problems the software is designed to solve, who the target audience is, and its relation to other work.

Please kindly let us know whether the above changes are needed, and what our next steps should be. Thank you very much.
