### 2024-07-09 #6507 @drvinceknight
> We thought we should wait for the reviews from @KBodolai to be completed before we start prepare the revisions.

@KBodolai would you be able to follow up on your review?

> Also, could you let us know how to upload the revisions?

If you make revisions to the paper and commit them to the paper repository that would be good. Once you have done that you can ping the reviewers here to let them know.

>
> In the meantime, we do not see a lot of changes requested by the reviewers. The two (potential) improvements we identify are:
>
> 1. Add missing DOIs for the references
> 2. Add a "statement of need" section that clearly states what problems the software is designed to solve, who the target audience is, and its relation to other work.

That looks right to me from @Otomisin's review.
