### 2024-07-09 a151c97
Added missing DOIs and updated references.
