+++ paper.bib
@@ -11,21 +11,21 @@
 }
 
 @article{Nesterov2017Random,
-author = {Nesterov, Yurii and Spokoiny, Vladimir},
-title = {Random Gradient-Free Minimization of Convex Functions},
-year = {2017},
-issue_date = {April 2017},
-publisher = {Springer-Verlag},
-address = {Berlin, Heidelberg},
-volume = {17},
-number = {2},
-issn = {1615-3375},
-doi = {10.1007/s10208-015-9296-2},
-journal = {Found. Comput. Math.},
-month = {apr},
-pages = {527–566},
-numpages = {40},
-keywords = {68Q25, Random methods, Stochastic optimization, Convex optimization, 90C25, Derivative-free methods, Complexity bounds, 0C47}
+  author = {Nesterov, Yurii and Spokoiny, Vladimir},
+  title = {Random Gradient-Free Minimization of Convex Functions},
+  year = {2017},
+  issue_date = {April 2017},
+  publisher = {Springer-Verlag},
+  address = {Berlin, Heidelberg},
+  volume = {17},
+  number = {2},
+  issn = {1615-3375},
+  doi = {10.1007/s10208-015-9296-2},
+  journal = {Found. Comput. Math.},
+  month = {apr},
+  pages = {527–566},
+  numpages = {40},
+  keywords = {68Q25, Random methods, Stochastic optimization, Convex optimization, 90C25, Derivative-free methods, Complexity bounds, 0C47}
 }
 
 
@@ -39,6 +39,7 @@ keywords = {68Q25, Random methods, Stochastic optimization, Convex optimization,
   timestamp = {2013-08-07T15:30:51.000+0200},
   title = {{Lipschitzian optimization without the Lipschitz constant}},
   type = {{Article}},
+  doi = {{10.1007/bf00941892}},
   volume = {{79}},
   year = {{1993}}
 }
@@ -66,11 +67,12 @@ journal = {Proceedings of the Fourteenth International Conference on Artificial 
 
 
 @inproceedings{Dragan2019Asset,
-author = {Komljenovic, Dragan and Messaoudi, Darragi and Cote, Alain and Gaha, Mohamed and Vouligny, Luc and Alarie, Stephane and Blancke, Olivier},
-year = {2019},
-month = {07},
-booktitle = {World Congress on Resilience, Reliability and Asset Management},
-title = {Asset Management in Electrical Utilities in the Context of Business and Operational Complexity},
+  author = {Komljenovic, Dragan and Messaoudi, Darragi and Cote, Alain and Gaha, Mohamed and Vouligny, Luc and Alarie, Stephane and Blancke, Olivier},
+  year = {2019},
+  month = {07},
+  booktitle = {World Congress on Resilience, Reliability and Asset Management},
+  title = {Asset Management in Electrical Utilities in the Context of Business and Operational Complexity},
+  doi = {10.1007/978-3-030-64228-0_4},
 }
 
 @inproceedings{golovin2017google,
@@ -145,6 +147,7 @@ title = {Asset Management in Electrical Utilities in the Context of Business and
   number={19},
   pages={1876--1902},
   year={2009},
+  doi = {10.1016/j.tcs.2009.01.016},
   publisher={Elsevier}
 }
 
@@ -200,6 +203,7 @@ title = {Asset Management in Electrical Utilities in the Context of Business and
   number={56},
   pages={153–195},
   year={2016},
+  doi = {10.1613/jair.4742},
   publisher={Elsevier}
 }
 
[…]
