### 2024-07-09 e11d7be
removed unused references and updated the paper.
