+++ paper.bib
@@ -10,39 +10,6 @@
   pages   = {1-52},
 }
 
-@article{Nesterov2017Random,
-  author = {Nesterov, Yurii and Spokoiny, Vladimir},
-  title = {Random Gradient-Free Minimization of Convex Functions},
-  year = {2017},
-  issue_date = {April 2017},
-  publisher = {Springer-Verlag},
-  address = {Berlin, Heidelberg},
-  volume = {17},
-  number = {2},
-  issn = {1615-3375},
-  doi = {10.1007/s10208-015-9296-2},
-  journal = {Found. Comput. Math.},
-  month = {apr},
-  pages = {527–566},
-  numpages = {40},
-  keywords = {68Q25, Random methods, Stochastic optimization, Convex optimization, 90C25, Derivative-free methods, Complexity bounds, 0C47}
-}
-
-
-@article{Jones1993Lipschitzian,
-  added-at = {2013-08-07T15:30:49.000+0200},
-  author = {Jones, David and Perttunen, Cary and Stuckman, Bruce},
-  journal = {{Journal of Optimization Theory and Applications}},
-  month = {{Oct}},
-  number = {{1}},
-  pages = {{157-181}},
-  timestamp = {2013-08-07T15:30:51.000+0200},
-  title = {{Lipschitzian optimization without the Lipschitz constant}},
-  type = {{Article}},
-  doi = {{10.1007/bf00941892}},
-  volume = {{79}},
-  year = {{1993}}
-}
 
 
 @InProceedings{ammar20derivative,
@@ -56,48 +23,6 @@
   series = 	 {Proceedings of Machine Learning Research},
   month = 	 {26--28 Aug},
   publisher =    {PMLR},
-}
-
-@article{Chu2011Contextual,
-title = {Contextual Bandits with Linear Payoff Functions},
-author = {Wei Chu and Lihong Li and Lev Reyzin and Robert Schapire},
-year ={2011},
-journal = {Proceedings of the Fourteenth International Conference on Artificial Intelligence and Statistics (AISTATS)},
-}
-
-
-@inproceedings{Dragan2019Asset,
-  author = {Komljenovic, Dragan and Messaoudi, Darragi and Cote, Alain and Gaha, Mohamed and Vouligny, Luc and Alarie, Stephane and Blancke, Olivier},
-  year = {2019},
-  month = {07},
-  booktitle = {World Congress on Resilience, Reliability and Asset Management},
-  title = {Asset Management in Electrical Utilities in the Context of Business and Operational Complexity},
-  doi = {10.1007/978-3-030-64228-0_4},
-}
-
-@inproceedings{golovin2017google,
-  title={Google vizier: A service for black-box optimization},
-  author={Golovin, Daniel and Solnik, Benjamin and Moitra, Subhodeep and Kochanski, Greg and Karro, John and Sculley, D},
-  booktitle={Proceedings of the 23rd ACM SIGKDD international conference on knowledge discovery and data mining},
-  pages={1487--1495},
-  year={2017}
-}
-
-
-
-@book{cesa2006prediction,
-  title={Prediction, Learning, and Games},
-  author={Cesa-Bianchi, N and Lugosi, G},
-  year={2006},
-  publisher={Cambridge University Press}
-}
-
-
-@book{lattimore2020bandit,
-  title={Bandit algorithms},
-  author={Lattimore, Tor and Szepesv{\'a}ri, Csaba},
-  year={2020},
-  publisher={Cambridge University Press}
 }
 
 
@@ -119,38 +44,8 @@ journal = {Proceedings of the Fourteenth International Conference on Artificial 
 }
 
 
-@article{maurer2009empirical,
-  title={Empirical Bernstein bounds and sample variance penalization},
[…]
