### 2024-07-09 52806dc
Corrected a bug for reference rendering
