### 2024-07-09 d724930
removed unused references and updated the paper.
