### 2024-07-09 #6507 @WilliamLwj
> > We thought we should wait for the reviews from @KBodolai to be completed before we start prepare the revisions.
>
> @KBodolai would you be able to follow up on your review?
>
> > Also, could you let us know how to upload the revisions?
>
> If you make revisions to the paper and commit them to the paper repository that would be good. Once you have done that you can ping the reviewers here to let them know.
>
> > In the meantime, we do not see a lot of changes requested by the reviewers. The two (potential) improvements we identify are:
> >
> > 1. Add missing DOIs for the references
> > 2. Add a "statement of need" section that clearly states what problems the software is designed to solve, who the target audience is, and its relation to other work.
>
> That looks right to me from @Otomisin's review.

Hi @drvinceknight, thanks for getting back to us so quickly. Regarding the two areas to improve,

1. We found that some references in the bib file are actually unused in the paper and removed them. For the remaining references, some of the papers do not have DOIs, for example, Derivative-free order-robust optimisation (Ammar, H. et al 2020) so we cannot add any DOIs to them. We have added DOIs to the rest of the papers.

2. This point confuses us because we have a "Statement of need" section, which is our section 2 in the paper. You can see that from the pdf generated in [editorialbot message](https://github.com/openjournals/joss-reviews/issues/6507#issuecomment-2007104519) on Mar 19. Therefore, we are quite confused by what @Otomisin mean when saying our paper does not have this section.

We have committed all of our changes to the paper and you can find the newest version in our [Github actions](https://github.com/WilliamLwj/PyXAB/actions/runs/9864232742).
