### 2024-08-01 #6507 @WilliamLwj
Hi @drvinceknight , are there any updates regarding the status of our paper? Kindly let us know, thank you!
