### 2024-08-10 #6507 @drvinceknight
@KBodolai would you be able to take a look please (to confirm if you're happy with the changes)?
