### 2024-08-25 WilliamLwj/PyXAB#37 @KBodolai
paper review
Hello, I think you've done a great job at context-setting, and a fantastic work of pulling all of these algorithms together in one single, consistent API. On top of that, the docs are delightful.

 a few minor things from the paper review, relating to openjournals/joss-reviews#6507:

1. Minor typo in the second paragraph of the statement of need:  ` ``arms".` is started with a double backtick and finishes with double quotes.
2. Just after that, I believe there's a missing "be": just before "infeasible" in the following statement `Therefore, directly applying multi-armed bandit algorithms to such problems would infeasible`
3. The Quick Example section in the Readme could be made a bit more specific and self-evident by including the specific imports used in it.
4. Love the real world example - if I were to add anything here, it would be a comparison with other algorithms commonly used for hyperparameter tuning (time, optimality)

Overall fantastic job, well done!
