### 2024-08-25 d987f88
Updated paper to address typos, updated README.md to include imports, as in Issue #7
