### 2024-08-25 #6507 @WilliamLwj
Hi @KBodolai, thank you for your review! We have had an update for the paper and the repository to address your comments.

> 1. Minor typo in the second paragraph of the statement of need:  ``arms". is started with a double backtick and finishes with double quotes.

Addressed this in the paper. Thank you! You can find the newest version of the paper [here](https://github.com/WilliamLwj/PyXAB/actions/runs/10549998565) (In the generated artifacts)

> 2. Just after that, I believe there's a missing "be": just before "infeasible" in the following statement Therefore, directly applying multi-armed bandit algorithms to such problems would infeasible

Same as the above. Thank you for finding these typos!

> 3. The Quick Example section in the Readme could be made a bit more specific and self-evident by including the specific imports used in it.

Thanks! We have updated the quick example in the README. There is only one thing to import which is the HOO algorithm. You can find the new README [here](https://github.com/WilliamLwj/PyXAB/tree/paper)

> 4. Love the real world example - if I were to add anything here, it would be a comparison with other algorithms commonly used for hyperparameter tuning (time, optimality)

Thank you! But just to clarify, are you referring to the real-life example in [our doc](https://pyxab.readthedocs.io/en/latest/getting_started/auto_examples/plot_VHCT_Landmine.html)? We can provide some comparisons with other hyper-parameter tuning algorithms, but we would like to know what algorithms you think are the most important to compare with. Thank you!

We would love to see your PR. Many thanks in advance!
