### 2024-08-25 WilliamLwj/PyXAB#37 @WilliamLwj
Hi, thank you for the review! We've updated the repository and the document. Please find the details [here](https://github.com/openjournals/joss-reviews/issues/6507#issuecomment-2308992265)
