### 2024-08-26 #6507 @KBodolai
Fantastic, very happy with the changes, thanks for being so quick!

I was referring to the docs example, yes. At the end is a simple example, so I'd probably just go with some baseline implementation in sklearn, such as  RandomizedSearchCV, you could also go "fancier" and compare it to other methods, but I feel like that would be more of a benchmark than an example, and in practice many people stick to these simple methods anyway!

I may not have time in the following couple of weeks - but I sugget we move on with the publication of the paper, since this last point is a minor, incremental improvement on an otherwise good documentation.

@drvinceknight , I think this completes my review, let me know if I need to do anything else.
