### 2024-09-10 219145f
Update CITATION.cff
