### 2024-09-10 7cf2859
Update CITATION.cff
