+++ CITATION.cff
@@ -1,4 +1,4 @@
-cff-version: 1.2.0
+cff-version: 1.1.0
 message: "If you use this software, please cite it as below."
 authors:
 - family-names: "Li"
@@ -14,6 +14,6 @@ authors:
   given-names: "Qifan"
   orcid: "https://orcid.org/0000-0000-0000-0000"
 title: "PyXAB - A Python Library for X-Armed Bandit and Online Blackbox Optimization Algorithms"
-version: 0.0.0
+version: 0.3.0
 date-released: 2023-03-07
 url: "https://arxiv.org/abs/2303.04030"
