### 2024-09-10 3e589b7
Update CITATION.cff
