### 2024-09-10 6baa302
Update CITATION.cff
