### 2024-09-10 8a78a23
Update CITATION.cff
