### 2024-09-10 #6507 @WilliamLwj
Hi @drvinceknight, yes I have made a tagged release and archived the release to Zenodo, the version number is `v0.3.0-paper`, and the DOI is [10.5281/zenodo.13743085](https://zenodo.org/doi/10.5281/zenodo.13743085)
