### 2024-09-24 #6507 @drvinceknight
> Hi @drvinceknight, we had some last minute changes on the paper, which is already pushed to our repository. (see [commit](https://github.com/WilliamLwj/PyXAB/commit/d10707ee3f0c43400dbc0e906144813cca036b7f)) Specifically, Prof. Jean Honorio (https://www.cs.purdue.edu/homes/jhonorio/) moved from Purdue University to The University of Melbourne in Australia earlier this year during the review period, and we have updated his affiliation information.
>
> Also, we tried to add one footnote saying that this work was done before Wenjie joined Amazon (required by Amazon), but that didn't show up in the generated pdf so we just decided to leave it like that.
>
> Please let us know whether this is ok and if you have any concerns/questions. Many thanks in advance.

Thanks for this, could you up date the release and the doi?
