### 2024-10-04 #6507 @danielskatz
Just from the error message, it looks like a table label might be multiply defined. I would search the .md file for that ID, or perhaps just "summary"
