### 2024-10-04 #6507 @WilliamLwj
Hi @danielskatz @drvinceknight, I just checked our .md file and I only found three places where I used the word "summary",

1. The section `Summary`
2. The sentence `In Table \ref{tab: summary}, we provide the comparison among some of the algorithms implemented in PyXAB...`
3. The label `\label{tab: summary}`

I'm actually baffled, where exactly is the table label multiply-defined? [Please find our paper .md file here](https://github.com/WilliamLwj/PyXAB/blob/paper/paper.md)

Thank you very much.
