### 2024-10-04 #6507 @danielskatz
The issue might be the space in the label - I would change both instances of `tab: summary` to `tab:summary` and see if that fixes the problem.
