### 2024-10-04 b640610
removed spaces in {tab: summary} for compilation of paper
