+++ paper.md
@@ -82,7 +82,7 @@ subset of algorithms, see e.g., [@azar2014online], [@bartlett2019simple]. The un
 deteriorates the situation. 
 
 
-In Table \ref{tab: summary}, we provide the comparison among some of the algorithms implemented in PyXAB, including 
+In Table \ref{tab:summary}, we provide the comparison among some of the algorithms implemented in PyXAB, including 
 \texttt{HOO} [@bubeck2011X], \texttt{DOO} [@Munos2011Optimistic], \texttt{StoSOO} [@Valko13Stochastic], \texttt{HCT} [@azar2014online], \texttt{POO} [@Grill2015Blackbox]
 \texttt{GPO} [@shang2019general], \texttt{SequOOL} [@bartlett2019simple],  \texttt{StroquOOL} [@bartlett2019simple],  \texttt{VROOM} [@ammar20derivative].
 and \texttt{VHCT} [@li2021optimumstatistical].
@@ -107,7 +107,7 @@ and \texttt{VHCT} [@li2021optimumstatistical].
         \texttt{VHCT}   & yes & yes &no\\
        \hline
     \end{tabular}
-    \label{tab: summary}
+    \label{tab:summary}
 \end{table}
 
 To remove the barriers for future research in this area, we have developed PyXAB, a Python library of the existing 
