### 2024-10-04 #6507 @WilliamLwj
> The issue might be the space in the label - I would change both instances of `tab: summary` to `tab:summary` and see if that fixes the problem.

Thank you @danielskatz! We have removed the spaces in both instances and published a new release after the change.
@drvinceknight Could you use [10.5281/zenodo.13835537](https://doi.org/10.5281/zenodo.13835537) to try again? Thank you!
