### 2024-10-04 #6507 @danielskatz
@WilliamLwj - JOSS doesn't require the paper to be in the repo, so that wasn't strictly needed, but I'll see how it looks...
