### 2024-10-21 #6507 @xuanxu
I'd say that the label for the table must be in the caption, see https://joss.readthedocs.io/en/latest/paper.html#tables-and-figures
