### 2024-10-21 e493b46
Put label inside the caption to update our paper.
