### 2024-10-21 #6507 @WilliamLwj
Hi @xuanxu @danielskatz, thank you and I have committed the change! See https://github.com/WilliamLwj/PyXAB/commit/e493b465d5e2fbca2dc4be635ee59b6adfaa8dd2. The new release is [10.5281/zenodo.13963754](https://doi.org/10.5281/zenodo.13963754)
