### 2024-10-23 #6507 @crvernon
:wave: @WilliamLwj - I just need you to address the following before I accept this for publication:

In your Zenodo archive:
- [ ] The author ordering in your Zenodo archive should match the order listed in your paper.  Please edit the metadata of your Zenodo archive to reflect this.  No new release necessary.
- [ ] The license in your Zenodo archive does not match the license used in your GitHub repository.  Please also edit this to match.

- [ ] LINE 36:  Something is a bit off with the phrasing here.  How about, "...problems would be infeasible and more complex algorithms that may be more appropriate have been developed which are designed for these types of problems (Grill...)"
- [ ] LINE 39:  Missing space between "StoS00" and the citation parenthesis
- [ ] LINE 53:  Citations should appear as "e.g., Azar et al. (2014), Bartlett et al. (2019)."
- [ ] LINE 57:  Missing comma after "Grill et al., 2015)"
- [ ] Your footnote number 3 cannot support the bibtex citation.  Simply type in "...found in Bubeck et al. (2011)" in this case
- [ ] LINE 66:  Remove extra text "Figure" in "Figure Figure 1."
- [ ] LINE 92:  Should be, "...the learner obtains one point from the algorithm, evaluates it on the objective, and returns the reward..." where "evaluate" and "return" now have "s" added to the end of each word.
- [ ] LINE 102:  No need for the "s" as the end of "documentations"
- [ ] LINE 103:  "numpy" should be spelled "NumPy"
- [ ] LINE 134:  chi should be a capital X if not using the correct symbol.  To maintain formatting of characters in your bib file, place them in curly brackets.
- [ ] LINE 145: The "b" in "bayesian" should be captialized

Let me know when these changes have been made.  Thanks!
