### 2024-10-24 0d156ca
Update paper.md according to https://github.com/openjournals/joss-reviews/issues/6507#issuecomment-2433325341
