### 2024-10-24 9b46bc8
Update paper.bib according to https://github.com/openjournals/joss-reviews/issues/6507#issuecomment-2433325341
