+++ paper.bib
@@ -181,7 +181,7 @@
 
 
 @article{Li2022Federated,
-  title={Federated X-armed Bandit}, 
+  title={Federated $\chi$-armed Bandit}, 
   volume={38}, 
   url={https://ojs.aaai.org/index.php/AAAI/article/view/29267}, 
   DOI={10.1609/aaai.v38i12.29267}, 
@@ -196,7 +196,7 @@
 
 
 @InProceedings{Li2024Personalized,
-  title = 	 {Personalized Federated {X}-armed Bandit},
+  title = 	 {Personalized Federated $\chi$-armed Bandit},
   author =       {Li, Wenjie and Song, Qifan and Honorio, Jean},
   booktitle = 	 {Proceedings of The 27th International Conference on Artificial Intelligence and Statistics},
   pages = 	 {37--45},
