### 2024-10-24 #6507 @crvernon
Thanks @WilliamLwj!  I just created a PR back into your `paper` branch for some last few fixes.  Let me know when this is  in and we will move forward with acceptance.

See https://github.com/WilliamLwj/PyXAB/pull/38
